backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
pydantic-settings==2.10.1
psycopg[binary]==3.2.9
psycopg-pool==3.2.4
PyJWT==2.10.1
